+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), although currently without version numbers added.
 
+## 2023-10-11
+
+### Changed
+- GAME_IDENTIFIER packet in fceux dump script
+
+
 ## 2023-10-08
 
 ### Changed